refactoring gui module
