+++ src/secondary_threads/mod.rs
@@ -1,3 +1,3 @@
 pub mod check_updates;
 pub mod parse_packets;
-pub mod write_report;
+pub mod write_report_file;
